DSR-156: significantly reduce HTML bloat on SitRep/flash-update URLs
A regrettable oversight led the HTML response to be far
larger than necessary, causing loads of unused data to be
present in the HTML response of each SitRep.

Changing how we filter the data leads to significantly
less data being cached in HTML responses — up to 100KB
in some cases (depends on what is in Contentful).