DSR-260: remove package-lock.json completely
We have both yarn.lock and package-lock.json but our
build tools are only utilizing Yarn at the present.

To reduce the number of false security advisories we
should delete our package-lock file as it doesn't stay
up to date and causes advisories that don't apply to
the actual stack.