DSR-194: put CTF/sorting code where it belongs
Originally the homepage and SitRepList shared queries/templates
but that's no longer the case. However the codebase was still
setup as if that were the case. This re-org is a better reflection
of what it currently does, moving CTF stuff into SitRepList and
leaving AppBar to control only the static content it holds.